+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.3 (2023-July-12)
+
+- Added example of BACnetStack_SendWhoIsWithLimits
+- Change timers settings. Discovery every 30 seconds, Read Property every 15 seconds, Print the database every 3 seconds
+
 ## 0.0.2 (2023-July-12)
 
 - Added comments and cleaned up the example project